Add more robust character encoding handling and rollback.
Some browsers might send CP1252 even though the page is UTF8. We handle this
in a similar way to rails by embedding a utf8 character in the form and using this to
detect how the browser encoded the data. We assume that if it wasn't utf8
that the browser probably sent cp1252, since this is a superset of iso-8859-1.